Prevent users from spamming the join button
This adds a server side timeout that prevents users from switching to a
different room within 5 seconds. Some regions (cough Asia cough) just
takes a while longer to join. As there is still a use case for rejoining
the underlying issue (that the room is destroyed before it is rejoined)
is not fixed.