Make git blame ignore the mass reformatting commit
Git 2.23+ (Released in August 2019) adds the ability to ignore certain
commits when using git blame. This commit adds a file that contains the
id of yesterday's mass reformatting commit and sets up git to always use
this file when running git blame and friends

Credit to https://akrabat.com/ignoring-revisions-with-git-blame/ for
this trick.